test: drain the two oversized test files into topic-scoped siblings
Ticket oversized-test-files-drain (audit-triage 2026-09-05). Whole
describe blocks moved with their harness intact, never regex-converted;
counts pinned equal before/after.

- document-store.test.ts (2287 lines, 106 tests) → save/load core stays;
  + document-store.workspace-registry.test.ts (listWorkspaces /
    createWorkspace / renameWorkspace — named -registry because the
    existing document-store.workspace.test.ts covers the tree-semantics
    topic, a different thing);
  + document-store.compact.test.ts (compactDocument + auto-compact +
    disposal);
  + document-store.lifecycle.test.ts (the route-translation deleteDocument
    helper + deleting + renameDocumentPath).
  Still 106 tests across the four files. The vi.mock + awaited-import
  harness is per-file by necessity; each copy is pruned to what its
  topics use.

- DaemonDocumentPage.test.tsx (1827 lines, 42 tests) → mount/connection/
  chrome/wiring stays;
  + DaemonDocumentPage.versions.test.tsx (manual save + history panel +
    markdown history reachability);
  + DaemonDocumentPage.branches.test.tsx (branch UI + MergeToast).
  Still 42 tests across the three files.
