+++ packages/mcp-server/src/server/store/document-store.test.ts
@@ -1,14 +1,7 @@
 import { mkdtemp, rm } from 'node:fs/promises'
 import { tmpdir } from 'node:os'
 import { join } from 'node:path'
-import {
-  DocumentHasDescendantsError,
-  DocumentMoveIntoSelfError,
-  isWorkspaceNotFoundError,
-  isWorkspaceSegmentTakenError,
-  WorkspaceNotFoundError,
-} from '@kamiazya/whiteboard-ports'
-import { LoroDoc, LoroMap } from 'loro-crdt'
+import { LoroDoc } from 'loro-crdt'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 
 // Swap DATA_DIR to a temp directory through vi.mock.
@@ -24,28 +17,8 @@ vi.mock('../config.js', () => ({
 }))
 
 // Use dynamic import so it runs after the mock is resolved.
-const {
-  saveDocument,
-  loadDocument,
-  listDocuments,
-  listWorkspaces,
-  compactDocument,
-  renameDocumentPath,
-  ConflictError,
-  getDocumentKind,
-  setDocumentSavedListener,
-} = await import('./document-store.js')
-const {
-  scheduleAutoCompact,
-  uninstallAutoCompact,
-  disposeAutoCompact,
-  _inFlightAutoCompactCountForTests,
-  _isDisposingAutoCompactForTests,
-} = await import('./auto-compact.js')
-const { captureLogsForTests } = await import('../log.js')
-const { getDefaultServerDeps } = await import('../../di/default-server-deps.js')
-const { wbDocumentDelete } = await import('@kamiazya/whiteboard-server-core')
-const { FileVersionStore } = await import('./version-store.js')
+const { saveDocument, loadDocument, listDocuments, renameDocumentPath, getDocumentKind } =
+  await import('./document-store.js')
 const { createIsolatedDb } = await import('./db/test-helpers.js')
 
 let handle: Awaited<ReturnType<typeof createIsolatedDb>>
@@ -877,1411 +850,5 @@ describe('path validation - self-describing error messages', () => {
     await expect(saveDocument('session1', 'valid-top/.bad', new LoroDoc())).rejects.toThrow(
       /"valid-top\/\.bad"/,
     )
-  })
-})
-
-describe('listWorkspaces', () => {
-  beforeEach(async () => {
-    tempDir = await mkdtemp(join(tmpdir(), 'whiteboard-test-'))
-    await setupIsolatedDb()
-  })
-
-  afterEach(async () => {
-    await teardownIsolatedDb()
-    await rm(tempDir, { recursive: true, force: true })
-  })
-
-  it('lists workspaces seeded via saveDocument', async () => {
-    await saveDocument('session-active', 'a', new LoroDoc())
-    await saveDocument('session-old', 'a', new LoroDoc())
-
-    const workspaces = await listWorkspaces()
-    const ids = workspaces.map((s) => s.workspaceId)
-    expect(ids).toContain('session-active')
-    expect(ids).toContain('session-old')
-  })
-
-  it('returns an empty array when no workspaces have been saved yet', async () => {
-    const workspaces = await listWorkspaces()
-    expect(workspaces).toHaveLength(0)
-  })
-
-  // listWorkspaces is now backed by the workspaces table, so the previous
-  // "non-directory DATA_DIR" corruption check no longer applies.
-})
-
-// ADR-0019: the daemon's own DocumentIndex (CacheCoherentDocumentIndex) is
-// the first implementation that PERSISTS and SERVES segment/displayName —
-// the shared ports conformance suite deliberately stays accept-and-ignore,
-// so this echo/validation/conflict/non-clobber coverage lives here.
-describe('createWorkspace — ADR-0019 identity (segment/displayName)', () => {
-  beforeEach(async () => {
-    tempDir = await mkdtemp(join(tmpdir(), 'whiteboard-ws-identity-test-'))
-    await setupIsolatedDb()
-  })
-
-  afterEach(async () => {
-    await teardownIsolatedDb()
-    await rm(tempDir, { recursive: true, force: true })
-  })
-
-  it('echoes stored segment and displayName back through listWorkspaces', async () => {
-    const deps = await getDefaultServerDeps()
-    await deps.documentIndex.createWorkspace({
-      workspaceId: 'ws-echo',
-      segment: 'team-notes',
-      displayName: 'Team notes',
-    })
-
-    const rows = await deps.documentIndex.listWorkspaces()
-    const row = rows.find((r) => r.workspaceId === 'ws-echo')
-    expect(row).toEqual({
-      workspaceId: 'ws-echo',
-      segment: 'team-notes',
-      displayName: 'Team notes',
-    })
-  })
-
-  it('a legacy workspace with no identity lists with the keys absent, not null', async () => {
-    const deps = await getDefaultServerDeps()
-    await deps.documentIndex.createWorkspace({ workspaceId: 'ws-legacy' })
-
-    const rows = await deps.documentIndex.listWorkspaces()
-    const row = rows.find((r) => r.workspaceId === 'ws-legacy')
-    expect(row).toEqual({ workspaceId: 'ws-legacy' })
-  })
-
-  it('a bare re-create (the wb_document_create createWorkspace:true path) does not clobber stored identity', async () => {
-    const deps = await getDefaultServerDeps()
-    await deps.documentIndex.createWorkspace({
-      workspaceId: 'ws-preserved',
-      segment: 'kept',
-      displayName: 'Kept',
-    })
-
-    // Every wbDocumentCreate({ createWorkspace: true }) call reaches this
-    // bare shape — the identity claimed above must survive it.
-    await deps.documentIndex.createWorkspace({ workspaceId: 'ws-preserved' })
-
-    const rows = await deps.documentIndex.listWorkspaces()
-    const row = rows.find((r) => r.workspaceId === 'ws-preserved')
-    expect(row).toEqual({ workspaceId: 'ws-preserved', segment: 'kept', displayName: 'Kept' })
-  })
-
-  it('re-creating with identical fields is idempotent', async () => {
-    const deps = await getDefaultServerDeps()
-    const input = { workspaceId: 'ws-idem', segment: 'idem', displayName: 'Idem' }
-    await deps.documentIndex.createWorkspace(input)
-    await expect(deps.documentIndex.createWorkspace(input)).resolves.toBeUndefined()
-
-    const rows = await deps.documentIndex.listWorkspaces()
-    expect(rows.find((r) => r.workspaceId === 'ws-idem')).toEqual(input)
-  })
-
-  it('a second workspace claiming an already-taken segment is refused, and neither row is left partial', async () => {
-    const deps = await getDefaultServerDeps()
-    await deps.documentIndex.createWorkspace({ workspaceId: 'ws-first', segment: 'contested' })
-
-    const err = await deps.documentIndex
-      .createWorkspace({ workspaceId: 'ws-second', segment: 'contested' })
-      .catch((e: unknown) => e)
-    expect(isWorkspaceSegmentTakenError(err)).toBe(true)
-
-    const rows = await deps.documentIndex.listWorkspaces()
-    expect(rows.find((r) => r.workspaceId === 'ws-second')).toBeUndefined()
-    expect(rows.find((r) => r.workspaceId === 'ws-first')).toEqual({
-      workspaceId: 'ws-first',
-      segment: 'contested',
-    })
-  })
-
-  // The inbound half of zod-schema-discipline: what listWorkspaces serves
-  // was validated on write, so the strict outbound workspaceSummarySchema
-  // never meets a poisoned row.
-  it('rejects a ULID-shaped segment at the boundary, before any row is written', async () => {
-    const deps = await getDefaultServerDeps()
-    await expect(
-      deps.documentIndex.createWorkspace({
-        workspaceId: 'ws-rejected',
-        segment: '01ARZ3NDEKTSV4RRFFQ69G5FAV',
-      }),
-    ).rejects.toThrow()
-
-    const rows = await deps.documentIndex.listWorkspaces()
-    expect(rows.find((r) => r.workspaceId === 'ws-rejected')).toBeUndefined()
-  })
-
-  it('rejects an empty displayName at the boundary, before any row is written', async () => {
-    const deps = await getDefaultServerDeps()
-    await expect(
-      deps.documentIndex.createWorkspace({ workspaceId: 'ws-rejected-2', displayName: '' }),
-    ).rejects.toThrow()
-
-    const rows = await deps.documentIndex.listWorkspaces()
-    expect(rows.find((r) => r.workspaceId === 'ws-rejected-2')).toBeUndefined()
-  })
-})
-
-// The daemon's registry is the one `renameWorkspace` implementation the
-// shared conformance suite does not run: its runners are the in-memory
-// double, the browser's IndexedDB index, and the tree index over an
-// in-memory registry. What is specific here is the mechanism — a single
-// UPDATE against the `workspaces_segment_unique` index — so the cases below
-// are about that, not about restating the contract.
-describe('renameWorkspace — the daemon registry', () => {
-  beforeEach(async () => {
-    tempDir = await mkdtemp(join(tmpdir(), 'whiteboard-ws-rename-test-'))
-    await setupIsolatedDb()
-  })
-
-  afterEach(async () => {
-    await teardownIsolatedDb()
-    await rm(tempDir, { recursive: true, force: true })
-  })
-
-  it('writes both layers and serves them back through listWorkspaces', async () => {
-    const deps = await getDefaultServerDeps()
-    await deps.documentIndex.createWorkspace({ workspaceId: 'ws-rename', segment: 'before' })
-
-    const renamed = await deps.documentIndex.renameWorkspace({
-      workspaceId: 'ws-rename',
-      segment: 'after',
-      displayName: 'After',
-    })
-
-    expect(renamed).toEqual({
-      workspaceId: 'ws-rename',
-      segment: 'after',
-      displayName: 'After',
-    })
-    const rows = await deps.documentIndex.listWorkspaces()
-    expect(rows.find((r) => r.workspaceId === 'ws-rename')).toEqual(renamed)
-    expect((await deps.documentIndex.resolveWorkspace('after'))?.workspaceId).toBe('ws-rename')
-    expect(await deps.documentIndex.resolveWorkspace('before')).toBeNull()
-  })
-
-  // The unique index refuses the UPDATE itself, so there is no window
-  // between a check and the write it authorises — and the row the refused
-  // statement targeted is untouched, rather than half renamed.
-  it('is refused by the unique index, leaving the row as it was', async () => {
-    const deps = await getDefaultServerDeps()
-    await deps.documentIndex.createWorkspace({ workspaceId: 'ws-holder', segment: 'contested' })
-    await deps.documentIndex.createWorkspace({
-      workspaceId: 'ws-asking',
-      segment: 'asking',
-      displayName: 'Asking',
-    })
-
-    const err = await deps.documentIndex
-      .renameWorkspace({ workspaceId: 'ws-asking', segment: 'contested', displayName: 'Renamed' })
-      .catch((e: unknown) => e)
-    expect(isWorkspaceSegmentTakenError(err)).toBe(true)
-
-    const rows = await deps.documentIndex.listWorkspaces()
-    // The displayName the same call carried must not have landed either: a
-    // refused rename is one operation refused, not a partial write.
-    expect(rows.find((r) => r.workspaceId === 'ws-asking')).toEqual({
-      workspaceId: 'ws-asking',
-      segment: 'asking',
-      displayName: 'Asking',
-    })
-  })
-
-  it('names a workspace that had none, leaving its address alone', async () => {
-    const deps = await getDefaultServerDeps()
-    await deps.documentIndex.createWorkspace({ workspaceId: 'ws-unnamed', segment: 'keep-me' })
-
-    const renamed = await deps.documentIndex.renameWorkspace({
-      workspaceId: 'ws-unnamed',
-      displayName: 'Now named',
-    })
-
-    expect(renamed).toEqual({
-      workspaceId: 'ws-unnamed',
-      segment: 'keep-me',
-      displayName: 'Now named',
-    })
-  })
-
-  // Zero rows updated is the only signal a single UPDATE gives, and it has
-  // to mean "no such workspace" rather than "done".
-  it('fails WorkspaceNotFoundError for a workspace the registry does not hold', async () => {
-    const deps = await getDefaultServerDeps()
-    await expect(
-      deps.documentIndex.renameWorkspace({ workspaceId: 'ws-absent', displayName: 'nobody' }),
-    ).rejects.toThrow(WorkspaceNotFoundError)
-  })
-
-  it('rejects a ULID-shaped segment at the boundary, before any row is written', async () => {
-    const deps = await getDefaultServerDeps()
-    await deps.documentIndex.createWorkspace({ workspaceId: 'ws-guarded', segment: 'fine' })
-
-    await expect(
-      deps.documentIndex.renameWorkspace({
-        workspaceId: 'ws-guarded',
-        segment: '01ARZ3NDEKTSV4RRFFQ69G5FAV',
-      }),
-    ).rejects.toThrow()
-
-    const rows = await deps.documentIndex.listWorkspaces()
-    expect(rows.find((r) => r.workspaceId === 'ws-guarded')).toEqual({
-      workspaceId: 'ws-guarded',
-      segment: 'fine',
-    })
-  })
-})
-
-describe('compactDocument', () => {
-  beforeEach(async () => {
-    tempDir = await mkdtemp(join(tmpdir(), 'whiteboard-compact-test-'))
-    await setupIsolatedDb()
-    const { mkdir } = await import('node:fs/promises')
-    await mkdir(join(tempDir, 'session1'), { recursive: true })
-  })
-
-  afterEach(async () => {
-    await teardownIsolatedDb()
-    await rm(tempDir, { recursive: true, force: true })
-  })
-
-  it('does not compact when no version exists', async () => {
-    const doc = new LoroDoc()
-    doc.getMovableList('elements').insert(0, 'x')
-    doc.commit()
-    await saveDocument('session1', 'test', doc)
-
-    const store = new FileVersionStore()
-    const result = await compactDocument('session1', 'test', store)
-    expect(result.compacted).toBe(false)
-    expect(result.reason).toBe('no-versions')
-  })
-
-  it('returns no-file when the .loro file is missing', async () => {
-    const store = new FileVersionStore()
-    const result = await compactDocument('session1', 'missing', store)
-    expect(result).toEqual({ compacted: false, beforeBytes: 0, afterBytes: 0, reason: 'no-file' })
-  })
-
-  // canvas blobs no longer share their parent directory with the session
-  // dir, so the previous "non-directory parent" stat failure no longer
-  // applies. Coverage of the corrupt-snapshot branch lives below.
-
-  it('treats an invalid workspace record as corruption instead of compacting over it', async () => {
-    const { getDb } = await import('./db/index.js')
-    const { LibsqlDocumentStore } = await import('./libsql/libsql-document-store.js')
-    const { chunkSnapshot } = await import('@kamiazya/whiteboard-ports')
-    const { _clearWorkspaceDocCacheForTests } = await import('./document-store.js')
-    const doc = new LoroDoc()
-    const store = new FileVersionStore()
-    await saveDocument('session1', 'broken', doc)
-    await store.save('session1', 'broken', doc, { auto: true })
-    const db = await getDb(tempDir)
-    // Corrupt the workspace record's snapshot rows directly — the
-    // beforeBytes read succeeds (the header is intact), but re-opening the
-    // live workspace document decodes the garbage bytes and must surface
-    // corruption rather than silently compacting an empty document over
-    // real data.
-    const libsqlStore = new LibsqlDocumentStore(db)
-    const { manifest, chunks } = chunkSnapshot(Buffer.from('not-a-loro-snapshot'), 1_000_000)
-    await libsqlStore.saveSnapshot({
-      docRef: { kind: 'workspace-tree', workspaceId: 'session1' },
-      manifest,
-      chunks,
-      frontier: new Uint8Array(),
-    })
-    _clearWorkspaceDocCacheForTests()
-
-    await expect(compactDocument('session1', 'broken', store)).rejects.toThrow()
-  })
-
-  it('compacts at a version cut point while keeping restore working', async () => {
-    const { LoroMap } = await import('loro-crdt')
-    const doc = new LoroDoc()
-    const list = doc.getMovableList('elements')
-    // Build a larger op log by repeatedly adding and updating elements.
-    for (let i = 0; i < 30; i++) {
-      const m = list.insertContainer(list.length, new LoroMap())
-      m.set('id', `elem-${i}`)
-      m.set('x', i)
-    }
-    doc.commit()
-    await saveDocument('session1', 'test', doc)
-
-    const store = new FileVersionStore()
-    const v = await store.save('session1', 'test', doc, { auto: true })
-
-    // Add more operations after the saved version.
-    for (let i = 0; i < 30; i++) {
-      const m = list.insertContainer(list.length, new LoroMap())
-      m.set('id', `extra-${i}`)
-    }
-    doc.commit()
-    await saveDocument('session1', 'test', doc, { overwrite: true })
-
-    const result = await compactDocument('session1', 'test', store)
-    expect(result.compacted).toBe(true)
-    expect(result.afterBytes).toBeLessThan(result.beforeBytes)
-
-    // The live state should still have all 60 elements after compaction.
-    const live = await loadDocument('session1', 'test')
-    expect(live.getMovableList('elements').length).toBe(60)
-
-    // Restoring the oldest version should still work at the cut point.
-    const past = await store.load('session1', v.id, live)
-    expect(past).not.toBeNull()
-    expect(past!.getMovableList('elements').length).toBe(30)
-
-    // Compaction prunes history, not state, so the frontier written for the
-    // shallow workspace record must still match the live workspace doc's
-    // current oplog version — later saves keep appending from it.
-    const { getDb } = await import('./db/index.js')
-    const { LibsqlDocumentStore } = await import('./libsql/libsql-document-store.js')
-    const { getWorkspaceDoc } = await import('./document-store.js')
-    const db = await getDb(tempDir)
-    const libsqlStore = new LibsqlDocumentStore(db)
-    const frontierResult = await libsqlStore.readFrontier({
-      docRef: { kind: 'workspace-tree', workspaceId: 'session1' },
-    })
-    const workspaceDoc = await getWorkspaceDoc('session1')
-    expect(frontierResult?.frontier).toEqual(workspaceDoc.oplogVersion().encode())
-  })
-
-  it('writes lastCompactedAt only on successful compaction', async () => {
-    const { LoroMap } = await import('loro-crdt')
-
-    async function readLastCompactedAt(_path: string): Promise<number | null> {
-      // Workspace-level meta (S4b/S7): compaction folds the workspace
-      // record, so there is one stamp per workspace, not per document.
-      const { openWorkspaceDocIfStored } = await import('./document-store.js')
-      const { readWorkspaceMeta } = await import('@kamiazya/whiteboard-loro-adapter')
-      const workspaceDoc = await openWorkspaceDocIfStored('session1')
-      if (workspaceDoc === null) return null
-      return readWorkspaceMeta(workspaceDoc).lastCompactedAt ?? null
-    }
-
-    // Case 1: no version → reason='no-versions' → lastCompactedAt stays null.
-    const empty = new LoroDoc()
-    empty.getMovableList('elements').insert(0, 'x')
-    empty.commit()
-    await saveDocument('session1', 'untouched', empty)
-    const noopStore = new FileVersionStore()
-    const noop = await compactDocument('session1', 'untouched', noopStore)
-    expect(noop.reason).toBe('no-versions')
-    expect(await readLastCompactedAt('untouched')).toBeNull()
-
-    // Case 2: version cut available → reason='ok' → lastCompactedAt is set.
-    const doc = new LoroDoc()
-    const list = doc.getMovableList('elements')
-    for (let i = 0; i < 30; i++) {
-      const m = list.insertContainer(list.length, new LoroMap())
-      m.set('id', `elem-${i}`)
-    }
-    doc.commit()
-    await saveDocument('session1', 'big', doc)
-    const store = new FileVersionStore()
-    await store.save('session1', 'big', doc, { auto: true })
-    for (let i = 0; i < 30; i++) {
-      const m = list.insertContainer(list.length, new LoroMap())
-      m.set('id', `extra-${i}`)
-    }
-    doc.commit()
-    await saveDocument('session1', 'big', doc, { overwrite: true })
-
-    const before = Date.now()
-    const result = await compactDocument('session1', 'big', store)
-    expect(result.compacted).toBe(true)
-    const after = Date.now()
-    const stamp = await readLastCompactedAt('big')
-    expect(stamp).not.toBeNull()
-    expect(stamp!).toBeGreaterThanOrEqual(before)
-    expect(stamp!).toBeLessThanOrEqual(after)
-
-    // Workspace-level mirror (dual-plane collapse S4b): compaction operates
-    // on the workspace record's oplog, so the shared timestamp lives on the
-    // workspace meta — and only successful compaction writes it, same as the
-    // row stamp ('untouched' compacted nothing above).
-    const { openWorkspaceDocIfStored } = await import('./document-store.js')
-    const { readWorkspaceMeta } = await import('@kamiazya/whiteboard-loro-adapter')
-    const workspaceDoc = await openWorkspaceDocIfStored('session1')
-    expect(workspaceDoc).not.toBeNull()
-    if (workspaceDoc === null) throw new Error('unreachable')
-    const wsStamp = readWorkspaceMeta(workspaceDoc).lastCompactedAt
-    expect(wsStamp).toBeGreaterThanOrEqual(before)
-    expect(wsStamp).toBeLessThanOrEqual(after)
-  })
-
-  // compactDocument reads the workspace record, exports the shallow
-  // snapshot from the live workspace doc, and writes it back — all under the
-  // workspace write lock, the same lock every content writer holds. A
-  // concurrent tool write therefore lands strictly before or strictly after
-  // the fold, and survives either way.
-  it('does not lose a concurrent tool write racing the compaction', async () => {
-    const { getDb } = await import('./db/index.js')
-    const { getDataDir } = await import('../config.js')
-    const { resolveDocumentIdAtPath } = await import('./document-store.js')
-    const { LibsqlDocumentStore } = await import('./libsql/libsql-document-store.js')
-    const { WorkspaceRoutedDocumentStore } = await import('./workspace-plane.js')
-    const { chunkSnapshot, reassembleSnapshot } = await import('@kamiazya/whiteboard-ports')
-
-    const doc = new LoroDoc()
-    const list = doc.getMovableList('elements')
-    for (let i = 0; i < 30; i++) {
-      const m = list.insertContainer(list.length, new LoroMap())
-      m.set('id', `elem-${i}`)
-    }
-    doc.commit()
-    await saveDocument('session1', 'test', doc)
-
-    const store = new FileVersionStore()
-    await store.save('session1', 'test', doc, { auto: true })
-
-    for (let i = 0; i < 30; i++) {
-      const m = list.insertContainer(list.length, new LoroMap())
-      m.set('id', `extra-${i}`)
-    }
-    doc.commit()
-    await saveDocument('session1', 'test', doc, { overwrite: true })
-
-    const db = await getDb(getDataDir())
-    const documentId = (await resolveDocumentIdAtPath('session1', 'test'))!
-
-    // Hold the cut lookup open so the tool write races into compaction's
-    // window instead of trivially serializing before it.
-    const delayedStore = new Proxy(store, {
-      get(target, prop, receiver) {
-        if (prop === 'earliestWorkspaceFrontiers') {
-          return async (workspaceId: string) => {
-            await new Promise((r) => setTimeout(r, 60))
-            return target.earliestWorkspaceFrontiers(workspaceId)
-          }
-        }
-        return Reflect.get(target, prop, receiver)
-      },
-    })
-
-    const compactPromise = compactDocument('session1', 'test', delayedStore)
-    await new Promise((r) => setTimeout(r, 20))
-    const routed = new WorkspaceRoutedDocumentStore(new LibsqlDocumentStore(db))
-    const docRef = { kind: 'document' as const, workspaceId: 'session1', documentId }
-    const existing = await routed.loadSnapshot({ docRef })
-    const toolDoc = new LoroDoc()
-    toolDoc.import(reassembleSnapshot(existing!.manifest, existing!.chunks))
-    toolDoc.getMap('root').set('agentMarker', 'concurrent-edit')
-    toolDoc.commit()
-    const chunked = chunkSnapshot(new Uint8Array(toolDoc.export({ mode: 'snapshot' })), 1_000_000)
-    const toolWrite = routed.saveSnapshot({
-      docRef,
-      manifest: chunked.manifest,
-      chunks: chunked.chunks,
-      frontier: new Uint8Array(toolDoc.oplogVersion().encode()),
-    })
-
-    await Promise.all([compactPromise, toolWrite])
-
-    const after = await loadDocument('session1', 'test')
-    expect(after.getMap('root').get('agentMarker')).toBe('concurrent-edit')
-    expect(after.getMovableList('elements').length).toBe(60)
-  })
-})
-
-// `document-store.deleteDocument` is gone: it was a second implementation of
-// `wb_document_delete`, and the HTTP route that called it is now an adapter
-// over the operation (ADR-0018). These cases are storage-level — blobs,
-// thumbnails, rows, the descendant refusal — so they stay at this layer and
-// drive the surviving implementation instead.
-//
-// The local helper is the ROUTE's translation, written once here rather than
-// at each call below: this surface addresses a document by path and answers
-// `false` for one that does not exist, where the operation addresses it by
-// the id the index assigned and throws. Everything else the cases assert is
-// unchanged, which is the point — the implementation moved, the behaviour
-// did not.
-async function deleteDocument(workspaceId: string, path: string): Promise<boolean> {
-  const deps = await getDefaultServerDeps()
-  // The tree index throws for an unknown WORKSPACE where the retired SQL
-  // index answered null; the route's translation treats both as absent.
-  const entry = await deps.documentIndex
-    .resolveDocument({ workspaceId, path })
-    .catch((err: unknown) => {
-      if (isWorkspaceNotFoundError(err)) return null
-      throw err
-    })
-  if (entry === null) return false
-  await wbDocumentDelete(deps, { workspaceId, documentId: entry.documentId })
-  return true
-}
-
-describe('deleting a document', () => {
-  beforeEach(async () => {
-    tempDir = await mkdtemp(join(tmpdir(), 'whiteboard-delete-test-'))
-    await setupIsolatedDb()
-    const { mkdir } = await import('node:fs/promises')
-    await mkdir(join(tempDir, 'session1'), { recursive: true })
-  })
-
-  afterEach(async () => {
-    await teardownIsolatedDb()
-    await rm(tempDir, { recursive: true, force: true })
-  })
-
-  // The MCP surface already refuses this, and for a stated reason: deletion
-  // is the operation with nothing to undo it, so the caller has to name what
-  // it is destroying. The HTTP surface deleting the same document silently
-  // strands every child under a prefix nothing owns.
-  it('refuses to delete a document that still has descendants', async () => {
-    await saveDocument('session1', 'a', new LoroDoc())
-    await saveDocument('session1', 'a/child', new LoroDoc())
-
-    await expect(deleteDocument('session1', 'a')).rejects.toThrow(DocumentHasDescendantsError)
-
-    const paths = (await listDocuments('session1')).map((c) => c.path).sort()
-    expect(paths).toEqual(['a', 'a/child'])
-  })
-
-  it('deletes a document whose name merely prefixes a sibling', async () => {
-    await saveDocument('session1', 'a', new LoroDoc())
-    await saveDocument('session1', 'a-sibling', new LoroDoc())
-
-    await expect(deleteDocument('session1', 'a')).resolves.toBe(true)
-
-    expect((await listDocuments('session1')).map((c) => c.path)).toEqual(['a-sibling'])
-  })
-
-  it('removes the tree entry, deletes branches/versions rows explicitly, and unlinks the version thumbnail PNGs, leaving the workspace row and a sibling canvas untouched', async () => {
-    const { getDb } = await import('./db/index.js')
-    const { createBranch } = await import('./branches-store.js')
-    const { stat } = await import('node:fs/promises')
-    const { LibsqlDocumentStore } = await import('./libsql/libsql-document-store.js')
-
-    const doc = new LoroDoc()
-    await saveDocument('session1', 'canvas-a', doc)
-    await saveDocument('session1', 'canvas-b', doc)
-    const store = new FileVersionStore()
-    const version = await store.save('session1', 'canvas-a', doc, { auto: true })
-    await store.saveThumbnail('session1', 'canvas-a', version.id, new Uint8Array([1, 2, 3]))
-    await createBranch('session1', 'canvas-a', { name: 'feature' })
-
-    const db = await getDb(tempDir)
-    const { resolveDocumentIdAtPath } = await import('./document-store.js')
-    const documentId = await resolveDocumentIdAtPath('session1', 'canvas-a')
-    if (documentId === null) throw new Error('document missing from the tree')
-    const libsqlStore = new LibsqlDocumentStore(db)
-
-    const thumbPath = join(tempDir, 'blobs', 'session1', 'versions', `${version.id}.png`)
-    await expect(stat(thumbPath)).resolves.toBeDefined()
-
-    await expect(deleteDocument('session1', 'canvas-a')).resolves.toBe(true)
-
-    const branchesAfter = await db
-      .selectFrom('branches')
-      .selectAll()
-      .where('documentId', '=', documentId)
-      .execute()
-    expect(branchesAfter).toEqual([])
-    const versionsAfter = await db
-      .selectFrom('versions')
-      .selectAll()
-      .where('documentId', '=', documentId)
-      .execute()
-    expect(versionsAfter).toEqual([])
-
-    // Content now lives in the workspace record: the tree node is gone and
-    // the bytes were EVACUATED into the trash, not destroyed.
-    const { DocumentStoreWorkspaceDocs } = await import('@kamiazya/whiteboard-workspace-index')
-    const { resolveWorkspaceDocumentById, readTrashEntries } = await import(
-      '@kamiazya/whiteboard-loro-adapter'
-    )
-    const stored = await new DocumentStoreWorkspaceDocs(libsqlStore).open('session1')
-    expect(stored).not.toBeNull()
-    expect(resolveWorkspaceDocumentById(stored!, documentId)).toBeNull()
-    expect(readTrashEntries(stored!).map((t) => t.documentId)).toContain(documentId)
-    await expect(stat(thumbPath)).rejects.toThrow()
-
-    const wsRow = await db
-      .selectFrom('workspaces')
-      .select(['id'])
-      .where('id', '=', 'session1')
-      .executeTakeFirst()
-    expect(wsRow).toBeDefined()
-    // The sibling lives in the tree (documents rows retired with the
-    // dual-plane collapse), so "untouched" is a listing fact.
-    expect((await listDocuments('session1')).map((c) => c.path)).toEqual(['canvas-b'])
-  })
-
-  // The defect this closes: wb_document_delete removed the index row and the
-  // Libsql bytes and stopped there, so a document an agent deleted left its
-  // thumbnails, its blob and a cached doc instance behind — while the same
-  // document deleted through the HTTP route did not. Both paths now run the
-  // same teardown, and this asserts on the FILES, not on the tool answering
-  // { deleted: true }, which it did throughout the whole defect.
-  it('leaves the same state when the delete comes through wb_document_delete as through the HTTP path', async () => {
-    const { getDb } = await import('./db/index.js')
-    const { stat } = await import('node:fs/promises')
-    const { wbDocumentDelete } = await import('@kamiazya/whiteboard-server-core')
-    const { LoroWorkspaceDocumentIndex } = await import('@kamiazya/whiteboard-workspace-index')
-    const { FsBlobStore } = await import('./fs/fs-blob-store.js')
-    const { LibsqlDocumentStore } = await import('./libsql/libsql-document-store.js')
-    const { peekDoc } = await import('./doc-cache.js')
-    const { cacheBackedWorkspaceDocs, documentTeardown, resolveDocumentIdAtPath } = await import(
-      './document-store.js'
-    )
-
-    const doc = new LoroDoc()
-    await saveDocument('session1', 'agent-deleted', doc)
-    const store = new FileVersionStore()
-    const version = await store.save('session1', 'agent-deleted', doc, { auto: true })
-    await store.saveThumbnail('session1', 'agent-deleted', version.id, new Uint8Array([1, 2, 3]))
-    // Populate the doc cache the way a read would, so eviction has something
-    // to evict — otherwise this half of the assertion passes vacuously.
-    // getDoc, not loadDocument: only the former goes through the LRU.
-    const { getDoc } = await import('./document-store.js')
-    await getDoc('session1', 'agent-deleted')
-    expect(peekDoc('session1', 'agent-deleted')).toBeDefined()
-
-    const db = await getDb(tempDir)
-    const documentId = await resolveDocumentIdAtPath('session1', 'agent-deleted')
-    expect(documentId).not.toBeNull()
-    if (documentId === null) throw new Error('unreachable')
-    const thumbPath = join(tempDir, 'blobs', 'session1', 'versions', `${version.id}.png`)
-    await expect(stat(thumbPath)).resolves.toBeDefined()
-
-    await wbDocumentDelete(
-      {
-        documentStore: new LibsqlDocumentStore(db),
-        blobStore: {} as never,
-        documentIndex: new LoroWorkspaceDocumentIndex(
-          cacheBackedWorkspaceDocs(),
-          new FsBlobStore(join(tempDir, 'blobs')),
-        ),
-        documentTeardown,
-      },
-      { workspaceId: 'session1', documentId },
-    )
-
-    await expect(stat(thumbPath)).rejects.toThrow()
-    expect(peekDoc('session1', 'agent-deleted')).toBeUndefined()
-    expect(await resolveDocumentIdAtPath('session1', 'agent-deleted')).toBeNull()
-  })
-
-  it('returns false for a missing canvas without throwing; deleting the same canvas twice returns true then false', async () => {
-    await expect(deleteDocument('session1', 'ghost')).resolves.toBe(false)
-
-    await saveDocument('session1', 'once', new LoroDoc())
-    await expect(deleteDocument('session1', 'once')).resolves.toBe(true)
-    await expect(deleteDocument('session1', 'once')).resolves.toBe(false)
-  })
-
-  it('deletes a canvas that never had an FS blob (nothing writes one post-collapse)', async () => {
-    // saveDocument never writes an FS blob — content lives entirely in the
-    // workspace record's Libsql-backed bytes.
-    await saveDocument('session1', 'row-only', new LoroDoc())
-
-    await expect(deleteDocument('session1', 'row-only')).resolves.toBe(true)
-    expect(
-      await (await import('./document-store.js')).resolveDocumentIdAtPath('session1', 'row-only'),
-    ).toBeNull()
-  })
-})
-
-describe('renameDocumentPath', () => {
-  beforeEach(async () => {
-    tempDir = await mkdtemp(join(tmpdir(), 'whiteboard-rename-test-'))
-    await setupIsolatedDb()
-    const { mkdir } = await import('node:fs/promises')
-    await mkdir(join(tempDir, 'session1'), { recursive: true })
-  })
-
-  afterEach(async () => {
-    await teardownIsolatedDb()
-    await rm(tempDir, { recursive: true, force: true })
-  })
-
-  it('moves only the path: branches/versions rows and the Libsql snapshot stay byte-identical and keyed to the same documentId', async () => {
-    const { getDb } = await import('./db/index.js')
-    const { createBranch, loadDocumentBranches } = await import('./branches-store.js')
-
-    const doc = new LoroDoc()
-    await saveDocument('session1', 'a', doc)
-    await createBranch('session1', 'a', { name: 'feature' })
-    const store = new FileVersionStore()
-    const version = await store.save('session1', 'a', doc, { auto: true })
-
-    const db = await getDb(tempDir)
-    const beforeId = await (await import('./document-store.js')).resolveDocumentIdAtPath(
-      'session1',
-      'a',
-    )
-    if (beforeId === null) throw new Error('document missing from the tree')
-    const before = { id: beforeId }
-    const documentId = before.id
-    const contentBefore = (await loadDocument('session1', 'a')).toJSON()
-
-    await expect(renameDocumentPath('session1', 'a', 'b')).resolves.toEqual({ documentId })
-
-    const list = await listDocuments('session1')
-    expect(list.map((c) => c.path)).toEqual(['b'])
-
-    const afterId = await (await import('./document-store.js')).resolveDocumentIdAtPath(
-      'session1',
-      'b',
-    )
-    if (afterId === null) throw new Error('document missing from the tree')
-    const after = { id: afterId }
-    expect(after.id).toBe(documentId)
-
-    const branchesAfter = await db
-      .selectFrom('branches')
-      .selectAll()
-      .where('documentId', '=', documentId)
-      .execute()
-    expect(branchesAfter.map((b) => b.name).sort()).toEqual(['feature', 'main'])
-
-    const versionsAfter = await db
-      .selectFrom('versions')
-      .selectAll()
-      .where('documentId', '=', documentId)
-      .execute()
-    expect(versionsAfter.map((v) => v.id)).toEqual([version.id])
-
-    // Content is untouched by the move — same documentId, same value.
-    expect((await loadDocument('session1', 'b')).toJSON()).toEqual(contentBefore)
-
-    // loadDocumentBranches also resolves under the new path.
-    const branches = await loadDocumentBranches('session1', 'b')
-    expect(branches.branches.map((b) => b.name).sort()).toEqual(['feature', 'main'])
-  })
-
-  it('returns null (never throws) for a missing source canvas', async () => {
-    await expect(renameDocumentPath('session1', 'ghost', 'somewhere')).resolves.toBeNull()
-  })
-
-  it('throws ConflictError for an already-taken target path and mutates neither canvas', async () => {
-    await saveDocument('session1', 'a', new LoroDoc())
-    await saveDocument('session1', 'b', new LoroDoc())
-
-    await expect(renameDocumentPath('session1', 'a', 'b')).rejects.toThrow(ConflictError)
-
-    const list = await listDocuments('session1')
-    expect(list.map((c) => c.path).sort()).toEqual(['a', 'b'])
-  })
-
-  it('throws the path validator error for an invalid target path', async () => {
-    await saveDocument('session1', 'a', new LoroDoc())
-    await expect(renameDocumentPath('session1', 'a', '../evil')).rejects.toThrow()
-  })
-
-  it('rename to the SAME path is a no-op success, returning the existing documentId', async () => {
-    await saveDocument('session1', 'a', new LoroDoc())
-    const { getDb } = await import('./db/index.js')
-    const _db = await getDb(tempDir)
-    const beforeId = await (await import('./document-store.js')).resolveDocumentIdAtPath(
-      'session1',
-      'a',
-    )
-    if (beforeId === null) throw new Error('document missing from the tree')
-    const before = { id: beforeId }
-
-    await expect(renameDocumentPath('session1', 'a', 'a')).resolves.toEqual({
-      documentId: before.id,
-    })
-
-    const list = await listDocuments('session1')
-    expect(list.map((c) => c.path)).toEqual(['a'])
-  })
-
-  // A path IS the hierarchy, so renaming one carries everything under it.
-  // Moving only the named row leaves its children addressed below a prefix
-  // no document owns — reachable by nothing the UI can show, and produced by
-  // the ordinary act of renaming a group.
-  it('carries every descendant with the renamed path', async () => {
-    await saveDocument('session1', 'design', new LoroDoc())
-    await saveDocument('session1', 'design/login', new LoroDoc())
-    await saveDocument('session1', 'design/deep/notes', new LoroDoc())
-    await saveDocument('session1', 'designs-elsewhere', new LoroDoc())
-
-    await renameDocumentPath('session1', 'design', 'product')
-
-    const paths = (await listDocuments('session1')).map((c) => c.path).sort()
-    expect(paths).toEqual(['designs-elsewhere', 'product', 'product/deep/notes', 'product/login'])
-  })
-
-  // The prefix test above would pass with a blind string replace; this one
-  // fails unless the rewrite is anchored at a segment boundary.
-  it('does not carry a sibling that merely shares the name as a prefix', async () => {
-    await saveDocument('session1', 'design', new LoroDoc())
-    await saveDocument('session1', 'design-system', new LoroDoc())
-
-    await renameDocumentPath('session1', 'design', 'product')
-
-    const paths = (await listDocuments('session1')).map((c) => c.path).sort()
-    expect(paths).toEqual(['design-system', 'product'])
-  })
-
-  // The one rule `planSubtreeMove` deliberately does NOT enforce, so each
-  // caller has to. Without it the depth-ordered write — correct for the
-  // upward move it was written for — is inverted, and the shallow row lands
-  // on a path its own descendant has not vacated yet.
-  it('refuses to move a document inside itself rather than raising a raw constraint error', async () => {
-    await saveDocument('session1', 'a', new LoroDoc())
-    await saveDocument('session1', 'a/x', new LoroDoc())
-
-    await expect(renameDocumentPath('session1', 'a', 'a/x')).rejects.toThrow(
-      DocumentMoveIntoSelfError,
-    )
-
-    const paths = (await listDocuments('session1')).map((c) => c.path).sort()
-    expect(paths).toEqual(['a', 'a/x'])
-  })
-
-  it('refuses to nest a document inside itself even when the destination is free', async () => {
-    await saveDocument('session1', 'a', new LoroDoc())
-    await saveDocument('session1', 'a/b', new LoroDoc())
-
-    await expect(renameDocumentPath('session1', 'a', 'a/nested')).rejects.toThrow(
-      DocumentMoveIntoSelfError,
-    )
-
-    const paths = (await listDocuments('session1')).map((c) => c.path).sort()
-    expect(paths).toEqual(['a', 'a/b'])
-  })
-
-  it('rejects a rename that would collide with an existing descendant path', async () => {
-    await saveDocument('session1', 'a', new LoroDoc())
-    await saveDocument('session1', 'a/x', new LoroDoc())
-    await saveDocument('session1', 'c/x', new LoroDoc())
-
-    // `a` -> `c` is free at the top, and still collides: `a/x` would land on
-    // the occupied `c/x`.
-    await expect(renameDocumentPath('session1', 'a', 'c')).rejects.toThrow(ConflictError)
-
-    const paths = (await listDocuments('session1')).map((c) => c.path).sort()
-    expect(paths).toEqual(['a', 'a/x', 'c/x'])
-  })
-
-  it('evicts every moved path from the cache, not only the two named ones', async () => {
-    const { peekDoc, clearCache } = await import('./doc-cache.js')
-    const { getDoc } = await import('./document-store.js')
-    clearCache()
-    await saveDocument('session1', 'a', new LoroDoc())
-    await saveDocument('session1', 'a/child', new LoroDoc())
-    await getDoc('session1', 'a/child')
-    expect(peekDoc('session1', 'a/child')).not.toBeUndefined()
-
-    await renameDocumentPath('session1', 'a', 'b')
-
-    // A stale instance cached under the old child path would be resurrected
-    // by the next read through it, shadowing the moved document.
-    expect(peekDoc('session1', 'a/child')).toBeUndefined()
-  })
-
-  it('evicts the old cache key so a subsequent getDoc under the old path misses the cache', async () => {
-    const { peekDoc, clearCache } = await import('./doc-cache.js')
-    const { getDoc } = await import('./document-store.js')
-    clearCache()
-    try {
-      await saveDocument('session1', 'a', new LoroDoc())
-      await getDoc('session1', 'a')
-      expect(peekDoc('session1', 'a')).toBeDefined()
-
-      await renameDocumentPath('session1', 'a', 'b')
-      expect(peekDoc('session1', 'a')).toBeUndefined()
-    } finally {
-      clearCache()
-    }
-  })
-
-  it('evicts a phantom doc-cache entry already sitting at the destination path, so the renamed content is not overwritten', async () => {
-    const { peekDoc, clearCache } = await import('./doc-cache.js')
-    const { getDoc } = await import('./document-store.js')
-    clearCache()
-    try {
-      // Write real content under 'a'.
-      const doc = new LoroDoc()
-      doc.getText('content').insert(0, 'real content')
-      doc.commit()
-      await saveDocument('session1', 'a', doc)
-
-      // Simulate a WS connect (or update route) against a not-yet-created
-      // path 'b': getDoc() lazily caches an empty in-memory doc for it
-      // even though there is no DB row yet.
-      await getDoc('session1', 'b')
-      expect(peekDoc('session1', 'b')).toBeDefined()
-
-      await renameDocumentPath('session1', 'a', 'b')
-
-      // The stale phantom doc must not still shadow the just-renamed
-      // canvas's real content at the destination path.
-      expect(peekDoc('session1', 'b')).toBeUndefined()
-
-      const reloaded = await getDoc('session1', 'b')
-      expect(reloaded.getText('content').toString()).toBe('real content')
-    } finally {
-      clearCache()
-    }
-  })
-})
-
-describe('auto-compact', () => {
-  beforeEach(async () => {
-    tempDir = await mkdtemp(join(tmpdir(), 'whiteboard-auto-compact-test-'))
-    await setupIsolatedDb()
-    const { mkdir } = await import('node:fs/promises')
-    await mkdir(join(tempDir, 'session1'), { recursive: true })
-  })
-
-  afterEach(async () => {
-    uninstallAutoCompact()
-    await disposeAutoCompact()
-    await teardownIsolatedDb()
-    await rm(tempDir, { recursive: true, force: true })
-  })
-
-  it('saveDocument notifies the registered document-saved listener', async () => {
-    const trigger = vi.fn<(workspaceId: string, path: string) => void>()
-    setDocumentSavedListener(trigger)
-    await saveDocument('session1', 'foo', new LoroDoc())
-    expect(trigger).toHaveBeenCalledTimes(1)
-    expect(trigger).toHaveBeenCalledWith('session1', 'foo')
-  })
-
-  it('scheduleAutoCompact debounces rapid triggers into a single compaction', async () => {
-    const { LoroMap } = await import('loro-crdt')
-
-    async function readLastCompactedAt(): Promise<number | null> {
-      const { openWorkspaceDocIfStored } = await import('./document-store.js')
-      const { readWorkspaceMeta } = await import('@kamiazya/whiteboard-loro-adapter')
-      const workspaceDoc = await openWorkspaceDocIfStored('session1')
-      if (workspaceDoc === null) return null
-      return readWorkspaceMeta(workspaceDoc).lastCompactedAt ?? null
-    }
-
-    // Build a canvas with a version cut + extra ops so compactDocument
-    // actually has work to do (otherwise the debounced firing would
-    // just return reason: 'no-versions' and lastCompactedAt stays null).
-    const doc = new LoroDoc()
-    const list = doc.getMovableList('elements')
-    for (let i = 0; i < 30; i++) {
-      const m = list.insertContainer(list.length, new LoroMap())
-      m.set('id', `e-${i}`)
-    }
-    doc.commit()
-    await saveDocument('session1', 'big', doc)
-    const store = new FileVersionStore()
-    await store.save('session1', 'big', doc, { auto: true })
-    for (let i = 0; i < 30; i++) {
-      const m = list.insertContainer(list.length, new LoroMap())
-      m.set('id', `x-${i}`)
-    }
-    doc.commit()
-    await saveDocument('session1', 'big', doc, { overwrite: true })
-
-    expect(await readLastCompactedAt()).toBeNull()
-
-    // Three rapid triggers within the debounce window must collapse into
-    // a single compactDocument run. Use a tiny debounce so the test stays fast.
-    scheduleAutoCompact('session1', 'big', store, { debounceMs: 50 })
-    scheduleAutoCompact('session1', 'big', store, { debounceMs: 50 })
-    scheduleAutoCompact('session1', 'big', store, { debounceMs: 50 })
-
-    // Nothing has fired yet.
-    expect(await readLastCompactedAt()).toBeNull()
-
-    // Wait past the debounce + the async compactDocument write. Poll instead of
-    // a fixed sleep so this does not flake on a slow CI runner.
-    const stamp = await vi.waitFor(
-      async () => {
-        const value = await readLastCompactedAt()
-        expect(value).not.toBeNull()
-        return value
-      },
-      { timeout: 2000 },
-    )
-    const settled = stamp!
-
-    // Further idle time without a new trigger must NOT re-compact. This half
-    // is an inherently bounded negative wait — keep it real rather than
-    // deleting the assertion.
-    await new Promise((r) => setTimeout(r, 150))
-    expect(await readLastCompactedAt()).toBe(settled)
-  })
-
-  it('keeps the cached doc coherent through an auto-compact — the next save does not clobber the compacted record', async () => {
-    // The legacy trap was a cached full doc re-exporting its entire history
-    // over a freshly shallowed record. The workspace record's compaction
-    // writes its OWN live doc's frontier, so the cache needs no eviction —
-    // this pins that a post-compact edit round-trips instead of resurrecting
-    // pre-compact bytes or losing the edit.
-    const { LoroMap } = await import('loro-crdt')
-    const { peekDoc, clearCache } = await import('./doc-cache.js')
-    const { getDoc } = await import('./document-store.js')
-
-    clearCache()
-
-    const doc = new LoroDoc()
-    const list = doc.getMovableList('elements')
-    for (let i = 0; i < 30; i++) {
-      const m = list.insertContainer(list.length, new LoroMap())
-      m.set('id', `e-${i}`)
-    }
-    doc.commit()
-    await saveDocument('session1', 'cached', doc)
-    const store = new FileVersionStore()
-    await store.save('session1', 'cached', doc, { auto: true })
-    for (let i = 0; i < 30; i++) {
-      const m = list.insertContainer(list.length, new LoroMap())
-      m.set('id', `x-${i}`)
-    }
-    doc.commit()
-    await saveDocument('session1', 'cached', doc, { overwrite: true })
-
-    await getDoc('session1', 'cached')
-    expect(peekDoc('session1', 'cached')).toBeDefined()
-
-    scheduleAutoCompact('session1', 'cached', store, { debounceMs: 50 })
-    await vi.waitFor(
-      async () => {
-        const { openWorkspaceDocIfStored } = await import('./document-store.js')
-        const { readWorkspaceMeta } = await import('@kamiazya/whiteboard-loro-adapter')
-        const workspaceDoc = await openWorkspaceDocIfStored('session1')
-        expect(
-          workspaceDoc && (readWorkspaceMeta(workspaceDoc).lastCompactedAt ?? null),
-        ).not.toBeNull()
-      },
-      { timeout: 2000 },
-    )
-
-    // Edit and save AFTER the compaction, then reload from stored bytes
-    // only: everything survives.
-    const live = await getDoc('session1', 'cached')
-    live.getMap('root').set('post-compact', 'edit')
-    live.commit()
-    await saveDocument('session1', 'cached', live, { overwrite: true })
-    const { _clearWorkspaceDocCacheForTests } = await import('./document-store.js')
-    clearCache()
-    _clearWorkspaceDocCacheForTests()
-    const reloaded = await loadDocument('session1', 'cached')
-    expect(reloaded.getMovableList('elements').length).toBe(60)
-    expect(reloaded.getMap('root').get('post-compact')).toBe('edit')
-  })
-})
-
-describe('auto-compact disposal', () => {
-  let disposedDb = false
-
-  beforeEach(async () => {
-    tempDir = await mkdtemp(join(tmpdir(), 'whiteboard-auto-compact-dispose-test-'))
-    await setupIsolatedDb()
-    const { mkdir } = await import('node:fs/promises')
-    await mkdir(join(tempDir, 'session1'), { recursive: true })
-    disposedDb = false
-  })
-
-  afterEach(async () => {
-    uninstallAutoCompact()
-    await disposeAutoCompact()
-    if (!disposedDb) {
-      await teardownIsolatedDb()
-    }
-    await rm(tempDir, { recursive: true, force: true })
-  })
-
-  async function buildCompactableCanvas(
-    path: string,
-  ): Promise<InstanceType<typeof FileVersionStore>> {
-    const { LoroMap } = await import('loro-crdt')
-    const doc = new LoroDoc()
-    const list = doc.getMovableList('elements')
-    for (let i = 0; i < 30; i++) {
-      const m = list.insertContainer(list.length, new LoroMap())
-      m.set('id', `e-${i}`)
-    }
-    doc.commit()
-    await saveDocument('session1', path, doc)
-    const store = new FileVersionStore()
-    await store.save('session1', path, doc, { auto: true })
-    for (let i = 0; i < 30; i++) {
-      const m = list.insertContainer(list.length, new LoroMap())
-      m.set('id', `x-${i}`)
-    }
-    doc.commit()
-    await saveDocument('session1', path, doc, { overwrite: true })
-    return store
-  }
-
-  // compactDocument normally settles fast enough (in-memory DB, tiny fixture)
-  // that polling for "in flight" would race the compaction to zero. Delay
-  // just the cut lookup — an await compactDocument makes early on — so tests
-  // can deterministically observe the in-flight window instead of depending
-  // on real-clock luck.
-  function withDelayedEarliestFrontiers(
-    store: InstanceType<typeof FileVersionStore>,
-    delayMs: number,
-  ): InstanceType<typeof FileVersionStore> {
-    return new Proxy(store, {
-      get(target, prop, receiver) {
-        if (prop === 'earliestWorkspaceFrontiers') {
-          return async (workspaceId: string) => {
-            await new Promise((r) => setTimeout(r, delayMs))
-            return target.earliestWorkspaceFrontiers(workspaceId)
-          }
-        }
-        return Reflect.get(target, prop, receiver)
-      },
-    })
-  }
-
-  it('cancels the pending debounce when the DB is disposed before it fires, instead of touching the destroyed driver', async () => {
-    const store = await buildCompactableCanvas('big')
-    const logs = captureLogsForTests('warning')
-
-    // Do NOT call uninstallAutoCompact() here — the point of this test
-    // is that DB disposal alone (without that manual call) must cancel the
-    // pending timer.
-    scheduleAutoCompact('session1', 'big', store, { debounceMs: 20 })
-    await teardownIsolatedDb()
-    disposedDb = true
-
-    // Wait past the debounce window. If the timer was not cancelled, its
-    // fired compactDocument call would hit the destroyed driver and log a
-    // 'failed' warning.
-    await new Promise((r) => setTimeout(r, 150))
-    logs.restore()
-
-    const autoCompactRecords = logs.records.filter((r) => r.scope === 'auto-compact')
-    expect(autoCompactRecords).toHaveLength(0)
-  })
-
-  it('disposeAutoCompact awaits an already-fired in-flight compaction before resolving', async () => {
-    const store = await buildCompactableCanvas('cached')
-
-    async function readLastCompactedAt(): Promise<number | null> {
-      const { openWorkspaceDocIfStored } = await import('./document-store.js')
-      const { readWorkspaceMeta } = await import('@kamiazya/whiteboard-loro-adapter')
-      const workspaceDoc = await openWorkspaceDocIfStored('session1')
-      if (workspaceDoc === null) return null
-      return readWorkspaceMeta(workspaceDoc).lastCompactedAt ?? null
-    }
-
-    scheduleAutoCompact('session1', 'cached', withDelayedEarliestFrontiers(store, 100), {
-      debounceMs: 1,
-    })
-    await vi.waitFor(
-      () => {
-        expect(_inFlightAutoCompactCountForTests()).toBeGreaterThan(0)
-      },
-      { timeout: 2000 },
-    )
-
-    await disposeAutoCompact()
-
-    expect(_inFlightAutoCompactCountForTests()).toBe(0)
-    expect(await readLastCompactedAt()).not.toBeNull()
-  })
-
-  it('disposeAutoCompact refuses a reschedule attempted while disposal is in progress, instead of racing a timer against the next clear pass', async () => {
-    const store = await buildCompactableCanvas('reentrant')
-
-    // Simulates loadDocument()'s legacy-migration path resuming mid-compaction
-    // and calling saveDocument(), which re-invokes the auto-compact trigger and
-    // attempts to schedule a fresh timer. The reschedule is gated on an
-    // explicit signal (not a wall-clock delay) so it fires deterministically
-    // once _isDisposingAutoCompactForTests() is confirmed true, rather than
-    // hoping a fixed sleep lands inside disposeAutoCompact's await window —
-    // that race is what made the previous version of this test flaky under
-    // load (CI run 29162596104).
-    let releaseReschedule: () => void = () => undefined
-    const rescheduleGate = new Promise<void>((resolve) => {
-      releaseReschedule = resolve
-    })
-    const rescheduleCallCount = { count: 0 }
-    const countingStore = new Proxy(store, {
-      get(target, prop, receiver) {
-        if (prop === 'earliestWorkspaceFrontiers') {
-          return async (workspaceId: string) => {
-            rescheduleCallCount.count += 1
-            return target.earliestWorkspaceFrontiers(workspaceId)
-          }
-        }
-        return Reflect.get(target, prop, receiver)
-      },
-    })
-    const reentrantStore = new Proxy(store, {
-      get(target, prop, receiver) {
-        if (prop === 'earliestWorkspaceFrontiers') {
-          return async (workspaceId: string) => {
-            await rescheduleGate
-            scheduleAutoCompact('session1', 'reentrant', countingStore, { debounceMs: 0 })
-            return target.earliestWorkspaceFrontiers(workspaceId)
-          }
-        }
-        return Reflect.get(target, prop, receiver)
-      },
-    })
-
-    scheduleAutoCompact('session1', 'reentrant', reentrantStore, { debounceMs: 1 })
-    await vi.waitFor(
-      () => {
-        expect(_inFlightAutoCompactCountForTests()).toBeGreaterThan(0)
-      },
-      { timeout: 2000 },
-    )
-
-    const disposePromise = disposeAutoCompact()
-    // Confirm disposal has actually begun (and is blocked awaiting the
-    // in-flight compaction above) before letting the reschedule proceed —
-    // this is the exact interleaving the original bug report depended on
-    // luck to hit.
-    await vi.waitFor(
-      () => {
-        expect(_isDisposingAutoCompactForTests()).toBe(true)
-      },
-      { timeout: 2000 },
-    )
-    releaseReschedule()
-    await disposePromise
-
-    expect(_inFlightAutoCompactCountForTests()).toBe(0)
-    // debounceMs: 0 still yields a macrotask tick before firing; give it a
-    // chance to fire if it were ever going to, then confirm it didn't.
-    await new Promise((r) => setTimeout(r, 20))
-    expect(rescheduleCallCount.count).toBe(0)
-  })
-
-  it('disposes through the real DB lifecycle (createIsolatedDb().dispose()) without spinning up a replacement connection for a re-entrant getDb() call', async () => {
-    const store = await buildCompactableCanvas('lifecycle')
-    const { getDb } = await import('./db/index.js')
-    let reentrantDb: Awaited<ReturnType<typeof getDb>> | null = null
-
-    const reentrantStore = new Proxy(store, {
-      get(target, prop, receiver) {
-        if (prop === 'earliestWorkspaceFrontiers') {
-          return async (workspaceId: string) => {
-            await new Promise((r) => setTimeout(r, 100))
-            // Mirrors a compaction resuming and touching the DB again
-            // (e.g. via loadDocument()) while teardown is draining hooks.
-            reentrantDb = await getDb(tempDir)
-            return target.earliestWorkspaceFrontiers(workspaceId)
-          }
-        }
-        return Reflect.get(target, prop, receiver)
-      },
-    })
-
-    scheduleAutoCompact('session1', 'lifecycle', reentrantStore, { debounceMs: 1 })
-    await vi.waitFor(
-      () => {
-        expect(_inFlightAutoCompactCountForTests()).toBeGreaterThan(0)
-      },
-      { timeout: 2000 },
-    )
-
-    const disposingDb = handle.db
-    // Exercise the actual DB lifecycle API (createIsolatedDb().dispose(),
-    // mirroring closeDb() in production) rather than calling
-    // disposeAutoCompact() directly, so the cache-removal-vs-hook-ordering
-    // fix in db/index.ts is covered from this store's perspective too.
-    await teardownIsolatedDb()
-    disposedDb = true
-
-    expect(reentrantDb).toBe(disposingDb)
-  })
-
-  it('is idempotent, and scheduleAutoCompact still works after a dispose', async () => {
-    const store = await buildCompactableCanvas('again')
-
-    async function readLastCompactedAt(): Promise<number | null> {
-      const { openWorkspaceDocIfStored } = await import('./document-store.js')
-      const { readWorkspaceMeta } = await import('@kamiazya/whiteboard-loro-adapter')
-      const workspaceDoc = await openWorkspaceDocIfStored('session1')
-      if (workspaceDoc === null) return null
-      return readWorkspaceMeta(workspaceDoc).lastCompactedAt ?? null
-    }
-
-    await disposeAutoCompact()
-    await disposeAutoCompact()
-
-    scheduleAutoCompact('session1', 'again', store, { debounceMs: 20 })
-    await vi.waitFor(
-      async () => {
-        expect(await readLastCompactedAt()).not.toBeNull()
-      },
-      { timeout: 2000 },
-    )
-  })
-
-  it('composes with uninstallAutoCompact() in either order without dropping in-flight work', async () => {
-    const store = await buildCompactableCanvas('composed')
-
-    async function readLastCompactedAt(): Promise<number | null> {
-      const { openWorkspaceDocIfStored } = await import('./document-store.js')
-      const { readWorkspaceMeta } = await import('@kamiazya/whiteboard-loro-adapter')
-      const workspaceDoc = await openWorkspaceDocIfStored('session1')
-      if (workspaceDoc === null) return null
-      return readWorkspaceMeta(workspaceDoc).lastCompactedAt ?? null
-    }
-
-    scheduleAutoCompact('session1', 'composed', withDelayedEarliestFrontiers(store, 100), {
-      debounceMs: 1,
-    })
-    await vi.waitFor(
-      () => {
-        expect(_inFlightAutoCompactCountForTests()).toBeGreaterThan(0)
-      },
-      { timeout: 2000 },
-    )
-
-    // uninstallAutoCompact() stays synchronous and timer-only: it must
-    // not swallow the in-flight compaction that is already running.
-    uninstallAutoCompact()
-    await disposeAutoCompact()
-
-    expect(await readLastCompactedAt()).not.toBeNull()
-  })
-
-  /**
-   * ADR-0020. Compaction folds a shallow snapshot out of the CACHED workspace
-   * document, so a cache that is behind writes a record missing another
-   * instance's ops — and the generation fence cannot catch it, because a
-   * writer that only APPENDED never touched the generation.
-   * `supersededDeltaCount` then drops exactly the delta that carried them.
-   */
-  it('does not fold away ops another instance appended while this cache was behind', async () => {
-    const { getDb } = await import('./db/index.js')
-    const { LibsqlDocumentStore } = await import('./libsql/libsql-document-store.js')
-    const { DocumentStoreWorkspaceDocs } = await import('@kamiazya/whiteboard-workspace-index')
-
-    // Enough history that a shallow snapshot is actually SMALLER than the
-    // record — otherwise compaction answers `no-gain` and returns without
-    // folding, and this case would pass against the very bug it is about.
-    const doc = new LoroDoc()
-    const elements = doc.getMovableList('elements')
-    for (let i = 0; i < 400; i += 1) {
-      elements.insert(i, `element-${i}-${'p'.repeat(200)}`)
-      doc.commit()
-    }
-    await saveDocument('session1', 'page', doc)
-
-    // Compaction refuses to cut history nothing pins, so the workspace needs
-    // a version before the fold is even attempted.
-    const store = new FileVersionStore()
-    await store.save('session1', 'page', await loadDocument('session1', 'page'), { auto: false })
-
-    // Another instance appends, straight to the store — this instance's
-    // cached workspace document knows nothing about it.
-    const docs = new DocumentStoreWorkspaceDocs(new LibsqlDocumentStore(await getDb(tempDir)))
-    const theirs = (await docs.open('session1')) as LoroDoc
-    theirs.getMap('meta').set('theirs', 'kept')
-    theirs.commit()
-    await docs.save('session1', theirs)
-
-    const result = await compactDocument('session1', 'page', store)
-    // Asserted, not assumed: a `no-gain` or `no-versions` answer means the
-    // fold never ran and everything below would pass for the wrong reason.
-    expect(result.reason).toBe('ok')
-    expect(result.compacted).toBe(true)
-
-    // The RECORD, not this process's cache: what survived the fold on disk.
-    const settled = (await docs.open('session1')) as LoroDoc
-    expect(settled.getMap('meta').get('theirs')).toBe('kept')
   })
 })
